db記入, controller @item_id
