+++ app/helpers/shipping_order_helper.rb
@@ -1,0 +1,2 @@
+module ShippingOrderHelper
+end
