+++ config/routes.rb
@@ -1,5 +1,16 @@
 Rails.application.routes.draw do
   devise_for :users
   root to: "items#index"
-  resources :items
+  resources :items do
+    resources :shipping_order, only: [:index, :create]
+  end
 end
+
+#Rails.application.routes.draw do
+  #devise_for :users
+  #root to: 'items#index'
+  #resources :items, only: %i[new create show edit update destroy] do
+    # 情報の紐付けはネストして紐付ける
+    #resources :buyer_history, only: %i[index create]
+  #end
+#end
