+++ db/migrate/20210411094756_create_orders.rb
@@ -1,7 +1,8 @@
 class CreateOrders < ActiveRecord::Migration[6.0]
   def change
     create_table :orders do |t|
-
+      t.references :user, foreign_key: true
+      t.references :item, foreign_key: true
       t.timestamps
     end
   end
