+++ spec/requests/shipping_order_request_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "ShippingOrders", type: :request do
+
+end
